orchestrator/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
httpx
python-dotenv
openai-chatkit
openai
autogen-ext[mcp]
autogen-agentchat
autogen-core
